Enable and fix footer
